+++ src/util-spm-bs2bm.c
@@ -40,7 +40,7 @@
  *                 characters that can't be inside the needle_len. So the skips can be
  *                 faster
  */
-void Bs2BmBadchars(const uint8_t *needle, uint32_t needle_len, uint8_t *badchars) {
+void Bs2BmBadchars(const uint8_t *needle, uint16_t needle_len, uint8_t *badchars) {
     uint32_t i;
     for (i = 0; i < ALPHABET_SIZE; i++)
         badchars[i] = 1;
@@ -61,7 +61,7 @@ void Bs2BmBadchars(const uint8_t *needle, uint32_t needle_len, uint8_t *badchars
  *                 characters that can't be inside the needle_len. So the skips can be
  *                 faster
  */
-void Bs2BmBadcharsNocase(const uint8_t *needle, uint32_t needle_len, uint8_t *badchars) {
+void Bs2BmBadcharsNocase(const uint8_t *needle, uint16_t needle_len, uint8_t *badchars) {
     uint32_t i;
     for (i = 0; i < ALPHABET_SIZE; i++)
         badchars[i] = 1;
@@ -88,7 +88,7 @@ void Bs2BmBadcharsNocase(const uint8_t *needle, uint32_t needle_len, uint8_t *ba
  *
  * \retval ptr to start of the match; NULL if no match
  */
-uint8_t * Bs2Bm(const uint8_t *haystack, uint32_t haystack_len, const uint8_t *needle, uint32_t needle_len, uint8_t badchars[])
+uint8_t * Bs2Bm(const uint8_t *haystack, uint32_t haystack_len, const uint8_t *needle, uint16_t needle_len, uint8_t badchars[])
 {
     const uint8_t *h, *n;
     const uint8_t *hmax = haystack + haystack_len;
@@ -137,7 +137,7 @@ uint8_t * Bs2Bm(const uint8_t *haystack, uint32_t haystack_len, const uint8_t *n
  *
  * \retval ptr to start of the match; NULL if no match
  */
-uint8_t *Bs2BmNocase(const uint8_t *haystack, uint32_t haystack_len, const uint8_t *needle, uint32_t needle_len, uint8_t badchars[])
+uint8_t *Bs2BmNocase(const uint8_t *haystack, uint32_t haystack_len, const uint8_t *needle, uint16_t needle_len, uint8_t badchars[])
 {
     const uint8_t *h, *n;
     const uint8_t *hmax = haystack + haystack_len;
