+++ src/util-spm-bs2bm.h
@@ -29,10 +29,10 @@
 
 #define ALPHABET_SIZE 256
 
-void Bs2BmBadchars(const uint8_t *, uint32_t, uint8_t *);
-void Bs2BmBadcharsNocase(const uint8_t *, uint32_t, uint8_t *);
-uint8_t * Bs2Bm(const uint8_t *, uint32_t, const uint8_t *, uint32_t, uint8_t []);
-uint8_t *Bs2BmNocase(const uint8_t *, uint32_t, const uint8_t *, uint32_t, uint8_t []);
+void Bs2BmBadchars(const uint8_t *, uint16_t, uint8_t *);
+void Bs2BmBadcharsNocase(const uint8_t *, uint16_t, uint8_t *);
+uint8_t * Bs2Bm(const uint8_t *, uint32_t, const uint8_t *, uint16_t, uint8_t []);
+uint8_t *Bs2BmNocase(const uint8_t *, uint32_t, const uint8_t *, uint16_t, uint8_t []);
 
 #endif /* __UTIL_SPM_BS2BM__ */
